requirements.txt
psycopg2-binary
kafka-python
google-generativeai
fastapi
uvicorn
websockets
